+++ src/main/java/nextstep/ladder/LadderApplication.java
@@ -1,9 +1,9 @@
 package nextstep.ladder;
 
-import nextstep.ladder.domain.Ladder;
-import nextstep.ladder.domain.LadderHeight;
-import nextstep.ladder.domain.Lines;
-import nextstep.ladder.domain.Participants;
+import nextstep.ladder.domain.ladder.Ladder;
+import nextstep.ladder.domain.ladder.LadderHeight;
+import nextstep.ladder.domain.ladder.Lines;
+import nextstep.ladder.domain.user.Participants;
 import nextstep.ladder.view.InputView;
 import nextstep.ladder.view.ResultView;
