+++ src/main/java/nextstep/ladder/domain/ladder/Lines.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.ladder;
 
 import nextstep.ladder.domain.generator.RandomBooleanGenerator;
