+++ src/main/java/nextstep/ladder/domain/user/Participants.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.user;
 
 import java.util.Arrays;
 import java.util.List;
